feat: refactored events & past events pages
